**Before**
- Average `/getallchatsofcharacter` Finish - `~1.37 s`
- Average Chats Load Finish - `~7.96 s`

**After**
- Average `/getallchatsofcharacter` Finish - `~942.50 ms`
- Average Chats Load Finish - `~3.48 s`

116 chats on a character, 4 attempts each.

---

I changed `/getchat` to use `JSON` instead of `json5` because `/savechat` already uses `JSON.stringify`.